Different pages in page now work, added basic onclicks
On the base page, "Create Restaurant" brings the page to a restaurant creation page. Then, the create restaurant button currently will bring the user to a success page. THIS NEEDS TO BE UPDATED so that correct id is given and to handle unsuccessful creations of restaurants. The back to login buttons should also work. Login still needs to be implemented, but wanted to pull this to have a template to work off of with different pages in one page.